:recycle: Refactor SignUp page wrapper
