+++ shared/common/src/pages/signUp/SignUpPages/SignUpButtonContainer/style.ts
