+++ shared/common/src/pages/signUp/SignUpPages/SignUpPage1/index.tsx
