+++ shared/common/src/pages/signUp/SignUpPages/SignUpPage2/index.tsx
