+++ shared/common/src/pages/signUp/SignUpPages/SignUpPage3/index.tsx
