+++ shared/common/src/pages/signUp/SignUpPages/SignUpPage3/style.ts
