+++ shared/common/src/pages/signUp/SignUpPages/SignUpScrollContainer/style.ts
