+++ shared/common/src/pages/signUp/SignUpPages/SignUpSuccess/index.stories.tsx
