+++ shared/common/src/pages/signUp/SignUpPages/SignUpSuccess/index.tsx
