+++ shared/common/src/pages/signUp/SignUpPages/SignUpSuccess/style.ts
