+++ shared/common/src/pages/signUp/SignUpPages/index.ts
